requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
fonttools==4.34.4
idna==3.3
influxdb==5.3.1
influxdb-client==1.31.0
kiwisolver==1.4.4
libvirt-python==8.6.0
matplotlib==3.5.2
msgpack==1.0.4
numpy==1.23.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
Rx==3.2.0
six==1.16.0
urllib3==1.26.11
